Add mode select, remove broken interface retrieve
Add ability to select a mouse mode (button configuration profile) using
'-s'. The Logitech G300s supports 3 modes, F3, F4 and F5.

Removing broken mouse_get_hid_interface_index() as it seems it's not the
way to determine interface index. Is there a way, if so I don't know it
yet so it's now hardcoded :(